+++ src/main.js
@@ -1,11 +1,11 @@
-import { getRandomString } from './utils/utils';
+import { getRandomString, isEscPressed } from './utils/utils';
 
-import SiteMenuView from './view/site-menu-view';
+import SiteMenuView from './view/site-menu/site-menu-view';
 import SiteFilterView from './view/site-filters/site-filters-view';
-import SiteSortView from './view/site-sort-view';
-import PointsListView from './view/site-list-view';
-import SitePointView from './view/site-point-view';
-import EditPointView from './view/site-edit-view';
+import SiteSortView from './view/site-sort/site-sort-view';
+import PointsListView from './view/site-list/site-list-view';
+import SitePointView from './view/site-points/site-point-view';
+import EditPointView from './view/site-edit/site-edit-view';
 
 import { renderElement, RenderPosition } from './utils/render';
 
@@ -22,8 +22,7 @@ const pointsListComponent = new PointsListView();
 renderElement(mainSort, pointsListComponent.element, RenderPosition.BEFOREEND);
 
 // GET THE DATA
-const url = 'https://16.ecmascript.pages.academy/big-trip/';
-const pointsUrl = `${url}points`;
+const url = 'https://16.ecmascript.pages.academy/big-trip/points';
 const fetchOptions = {
   method: 'GET',
   headers: {
@@ -47,7 +46,7 @@ const renderPoints = (points) => {
       };
 
       const onEscKeyDown = (evt) => {
-        if (evt.key === 'Escape' || evt.key === 'Esc') {
+        if (isEscPressed(evt)) {
           evt.preventDefault();
           replaceEditPointToPoint();
           document.removeEventListener('keydown', onEscKeyDown);
@@ -75,7 +74,7 @@ const renderPoints = (points) => {
   );
 };
 
-fetch(pointsUrl, fetchOptions)
+fetch(url, fetchOptions)
   .then((response) => response.json())
   .then((points) => renderPoints(points));
