+++ src/view/site-list/site-list.tpl.js
@@ -1,0 +1,4 @@
+export const createPointListTemplate = () => (
+  `<ul class="trip-events__list">
+  </ul>`
+);
