+++ src/view/site-sort/site-sort.tpl.js
@@ -1,6 +1,4 @@
-import { createElement } from '../utils/render';
-
-const createSiteSortTemplate = () => (
+export const createSiteSortTemplate = () => (
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
     <div class="trip-sort__item  trip-sort__item--day">
       <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked>
@@ -28,23 +26,3 @@ const createSiteSortTemplate = () => (
     </div>
   </form>`
 );
-
-export default class SiteSortView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  get template() {
-    return createSiteSortTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
-  }
-}
